refactor(rendering): extract 4k stage functions
